@@ -18,6 +18,7 @@
     "@types/node": "20.6.3",
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
+    "@uploadthing/react": "^5.6.2",
     "autoprefixer": "10.4.16",
     "axios": "^1.5.1",
     "class-variance-authority": "^0.7.0",
@@ -25,8 +26,8 @@
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.12",
     "lucide-react": "^0.279.0",
-    "next": "13.4.12",
-    "postcss": "8.4.30",
+    "next": "^13.5.5",
+    "postcss": "^8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.46.2",
@@ -35,6 +36,7 @@
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "5.2.2",
+    "uploadthing": "^5.7.1",
     "zod": "^3.22.2"
   },
   "devDependencies": {
