@@ -8,6 +8,7 @@ import { Pencil } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import toast from "react-hot-toast";
+import { Course } from "@prisma/client";
 
 import {
   Form,
@@ -15,26 +16,25 @@ import {
   FormField,
   FormItem,
 } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
+import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
+import { Textarea } from "@/components/ui/textarea";
 
-interface TitleFormProps {
-  initialData: {
-    title: string;
-  };
+interface DescriptionFormProps {
+  initialData: Course;
   courseId: string;
 };
 
 const fromSchema = z.object({
-  title: z.string().min(1, {
-    message: "Title is required.",
+  description: z.string().min(1, {
+    message: "Description is required.",
   }),
 });
 
-export const TitleForm = ({
+export const DescriptionForm = ({
   initialData,
   courseId,
-}: TitleFormProps) => {
+}: DescriptionFormProps) => {
   const [isEditing, setIsEditing] = useState(false);
   const toggleEdit = () => setIsEditing((current) => !current);
 
@@ -42,7 +42,9 @@ export const TitleForm = ({
 
   const form = useForm<z.infer<typeof fromSchema>>({
     resolver: zodResolver(fromSchema),
-    defaultValues: initialData,
+    defaultValues: {
+      description: initialData.description || "",
+    },
   });
 
   const { isSubmitting, isValid } = form.formState;
@@ -50,7 +52,7 @@ export const TitleForm = ({
   const onSubmit = async (values: z.infer<typeof fromSchema>) => {
     try {
       await axios.patch(`/api/courses/${courseId}`, values);
-      toast.success("Course title updated.");
+      toast.success("Course description updated.");
       toggleEdit();
       router.refresh();
     } catch {
@@ -61,21 +63,24 @@ export const TitleForm = ({
   return (
     <div className="mt-6 border bg-slate-100 rounded-md p-4">
       <div className="font-medium flex items-center justify-between">
-        Course Title
+        Course Description
         <Button onClick={toggleEdit} variant="ghost">
           {isEditing ? (
             <>Cancel</>
             ): (
               <>
               <Pencil className="h-4 w-4 mr-2" />
-              Edit Title
+              Edit description
             </>
           )}
         </Button>
       </div>
       {!isEditing && (
-        <p className="text-sm mt-2">
-          {initialData.title}
+        <p className={cn(
+          "text-sm mt-2",
+          !initialData && "text-slate-500 italic"
+        )}>
+          {initialData.description || "No description provided."}
         </p>
       )}
       {isEditing && (
@@ -86,13 +91,13 @@ export const TitleForm = ({
           >
             <FormField
               control={form.control}
-              name="title"
+              name="description"
               render={({ field }) => (
                 <FormItem>
                   <FormControl>
-                    <Input 
+                    <Textarea 
                       disabled={isSubmitting}
-                      placeholder="E.g. 'Adavanced web development'"
+                      placeholder="E.g. 'This course is about...'"
                       {...field}
                     />
                   </FormControl>
